separate WebviewPanel from base webview controller
